new_energy_project_management/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Login
python-dotenv
reportlab
pandas
openpyxl